Refactor genesis ingest functions
