+++ genesis-utils/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "solana-genesis-utils"
+version = "1.7.0"
+description = "Solana Genesis Utils"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-download-utils"
+edition = "2018"
+
+[dependencies]
+solana-sdk = { path = "../sdk", version = "=1.7.0" }
+solana-download-utils = { path = "../download-utils", version = "=1.7.0" }
+solana-runtime = { path = "../runtime", version = "=1.7.0" }
+
+[lib]
+crate-type = ["lib"]
+name = "solana_genesis_utils"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
