+++ validator/Cargo.toml
@@ -34,6 +34,7 @@ solana-client = { path = "../client", version = "=1.7.0" }
 solana-core = { path = "../core", version = "=1.7.0" }
 solana-download-utils = { path = "../download-utils", version = "=1.7.0" }
 solana-faucet = { path = "../faucet", version = "=1.7.0" }
+solana-genesis-utils = { path = "../genesis-utils", version = "=1.7.0" }
 solana-ledger = { path = "../ledger", version = "=1.7.0" }
 solana-logger = { path = "../logger", version = "=1.7.0" }
 solana-metrics = { path = "../metrics", version = "=1.7.0" }
